AllocationTimeline and AllocationQueueHeader refactored and tested
